requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
openpyxl>=3.1.0
numpy>=1.24.0
python-dotenv>=1.0.0